+++ src/sections/Learn/Kanvas-Labs/courseData.js
@@ -21,7 +21,7 @@ export const data = [
         title: "Working with Meshery and Istio",
         info: "Learn how to run Meshery, install Istio and deploy a sample app",
         embed: "istio-meshery-adapter",
-        link: "/learn/service-mesh-labs/working-with-meshery-and-istio",
+        link: "/learn/kanvas-labs/working-with-meshery-and-istio",
         cover: istioLogo
       },
       {
@@ -29,7 +29,7 @@ export const data = [
         title: "Working with Meshery and Linkerd",
         info: "Learn how to run Meshery, install Linkerd and deploy a sample app",
         embed: "linkerd-meshery-adapter",
-        link: "/learn/service-mesh-labs/working-with-meshery-and-linkerd",
+        link: "/learn/kanvas-labs/working-with-meshery-and-linkerd",
         cover: linkerdLogo
       },
       {
@@ -37,7 +37,7 @@ export const data = [
         title: "Working with Meshery and Consul",
         info: "Learn how to run Meshery, install Consul and deploy a sample app",
         embed: "consul-meshery-adapter",
-        link: "/learn/service-mesh-labs/working-with-meshery-and-consul",
+        link: "/learn/kanvas-labs/working-with-meshery-and-consul",
         cover: consulLogo
       },
       {
@@ -45,7 +45,7 @@ export const data = [
         title: "Working with Meshery and Kuma",
         info: "Learn how to run Meshery, install Kuma and deploy a sample app",
         embed: "kuma-meshery-adapter",
-        link: "/learn/service-mesh-labs/working-with-meshery-and-kuma",
+        link: "/learn/kanvas-labs/working-with-meshery-and-kuma",
         cover: kumaLogo
       },
       {
@@ -53,7 +53,7 @@ export const data = [
         title: "Working with Meshery and NSM",
         info: "Learn how to run Meshery, install NSM and deploy a sample app",
         embed: "nsm-meshery-adapter",
-        link: "/learn/service-mesh-labs/working-with-meshery-and-nsm",
+        link: "/learn/kanvas-labs/working-with-meshery-and-nsm",
         cover: nsmLogo
       },
       {
@@ -61,7 +61,7 @@ export const data = [
         title: "Working with Meshery and Traefik",
         info: "Learn how to run Meshery, install Traefik and deploy a sample app",
         embed: "traefik-meshery-adapter",
-        link: "/learn/service-mesh-labs/working-with-meshery-and-traefik",
+        link: "/learn/kanvas-labs/working-with-meshery-and-traefik",
         cover: traefikLogo
       },
       {
@@ -69,7 +69,7 @@ export const data = [
         title: "Working with Meshery and Open Service Mesh",
         info: "Learn how to run Meshery, install Open Service Mesh and deploy a sample app",
         embed: "osm-meshery-adapter",
-        link: "/learn/service-mesh-labs/working-with-meshery-and-open-service-mesh",
+        link: "/learn/kanvas-labs/working-with-meshery-and-open-service-mesh",
         cover: osmLogo
       }
     ]
@@ -85,7 +85,7 @@ export const data = [
         title: "Running Performance Tests",
         info: "Learn how to do performance testing with Meshery",
         embed: "running-performance-tests",
-        link: "/learn/service-mesh-labs/running-performance-tests",
+        link: "/learn/kanvas-labs/running-performance-tests",
         cover: smpLogo
       },
       {
@@ -93,7 +93,7 @@ export const data = [
         title: "Traffic Splitting with Meshery",
         info: "Traffic Management with Meshery",
         embed: "traffic-splitting",
-        link: "/learn/service-mesh-labs/traffic-splitting-with-meshery",
+        link: "/learn/kanvas-labs/traffic-splitting-with-meshery",
         cover: smiLogo
       }
     ]
